+++ src/importers/AmicaImporter.ts
@@ -2,16 +2,13 @@ import fetch, { Response } from 'node-fetch';
 import { startOfWeek, addWeeks, format, parse } from 'date-fns';
 
 import AbstractImporter from './AbstractImporter';
-import {
-  deleteMenusForRestaurantForDate,
-  createMenu,
-  CreateMenuParams,
-} from '../models/menu';
-import { CreateMenuItemParams, MenuItemType } from '../models/menuItem';
+import { CreateMenuParams } from '../menu/interfaces';
+import { deleteMenusForRestaurantForDate, createMenu } from '../menu/services';
+import { CreateMenuItemParams, MenuItemType } from '../menuItem/interfaces';
 import {
   CreateMenuItemComponentParams,
   MenuItemComponentType,
-} from '../models/menuItemComponent';
+} from '../menuItemComponent/interfaces';
 import { normalizeImportedString } from '../utils/normalize';
 
 interface SetMenu {
