+++ src/importers/SodexoImporter.ts
@@ -5,17 +5,14 @@ import fetch, { Response } from 'node-fetch';
 import { startOfWeek, addWeeks, format, addDays } from 'date-fns';
 
 import AbstractImporter from './AbstractImporter';
-import {
-  deleteMenusForRestaurantForDate,
-  createMenu,
-  CreateMenuParams,
-} from '../models/menu';
-import { CreateMenuItemParams, MenuItemType } from '../models/menuItem';
+import { CreateMenuParams } from '../menu/interfaces';
+import { deleteMenusForRestaurantForDate, createMenu } from '../menu/services';
+import { CreateMenuItemParams, MenuItemType } from '../menuItem/interfaces';
 import {
   CreateMenuItemComponentParams,
   MenuItemComponentType,
-} from '../models/menuItemComponent';
-import { ImportDetails } from '../models/importDetails';
+} from '../menuItemComponent/interfaces';
+import { ImportDetails } from '../importDetails/interfaces';
 import { normalizeImportedString } from '../utils/normalize';
 
 type TitleKey = 'title_fi' | 'title_en';
