+++ src/menu/services.ts
@@ -1,62 +1,9 @@
-// eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-// @ts-ignore
-import { knexPaginator as paginate } from 'apollo-cursor-pagination';
-import { gql, UserInputError } from 'apollo-server-express';
 import Knex from 'knex';
 import subWeeks from 'date-fns/sub_weeks';
 
-import { Context, Pagination } from '../utils/graphql';
-import { normalizeDatabaseData } from '../utils/normalize';
-import {
-  Restaurant,
-  RestaurantSearchConditions,
-  getRestaurant,
-} from './restaurant';
-import {
-  MenuItem,
-  CreateMenuItemParams,
-  getMenuItemsForMenu,
-  createMenuItem,
-} from './menuItem';
-
-export interface Menu {
-  id: number;
-  date: Date;
-  language: string;
-  created_at: Date;
-  updated_at: Date;
-  restaurant_id: number;
-  menu_items?: MenuItem[];
-  restaurant?: Restaurant;
-}
-
-export interface CreateMenuParams {
-  date: Date;
-  language: string;
-  restaurant_id: number;
-  menu_items?: CreateMenuItemParams[];
-}
-
-export interface MenuSearchConditions {
-  date?: string;
-  language?: string;
-  restaurant_id?: number;
-}
-
-export interface MenuSearchParams {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  columns: (string | Knex.Raw<any>)[];
-  conditions: MenuSearchConditions;
-  restaurantConditions: RestaurantSearchConditions;
-  order: string;
-}
-
-export const defaultSearchParams: MenuSearchParams = {
-  columns: ['menus.*'],
-  conditions: {},
-  restaurantConditions: { enabled: true },
-  order: 'restaurants.id ASC',
-};
+import { Menu, MenuSearchParams, CreateMenuParams } from './interfaces';
+import { getRestaurant } from '../restaurant/services';
+import { getMenuItemsForMenu, createMenuItem } from '../menuItem/services';
 
 export const getMenu = async (
   db: Knex,
@@ -210,120 +157,3 @@ export const deleteMenusOlderThan = async (
     .where('date', '<', subWeeks(Date(), weeks))
     .delete()
     .catch((): number => 0);
-
-export const menuTypeDefs = gql`
-  type Menu {
-    id: Int!
-    language: String!
-    date: Date!
-    createdAt: Date!
-    updatedAt: Date!
-    restaurant: Restaurant!
-    menuItems: [MenuItem]!
-  }
-
-  type MenuConnection {
-    pageInfo: PageInfo!
-    edges: [MenuEdge]!
-    totalCount: Int!
-  }
-
-  type MenuEdge {
-    cursor: String!
-    node: Menu!
-  }
-
-  extend type Query {
-    menu(id: Int!): Menu
-    menus(
-      first: Int
-      last: Int
-      before: String
-      after: String
-      orderBy: String
-      orderDirection: OrderDirection
-    ): MenuConnection!
-  }
-`;
-
-export const menuResolvers = {
-  Query: {
-    menu: async (
-      _: undefined,
-      { id }: { id: number },
-      { db }: Context,
-    ): Promise<object | undefined> => {
-      const data = await db<Menu>('menus').where('id', id);
-
-      if (data) {
-        return normalizeDatabaseData(data[0]);
-      }
-    },
-    menus: async (
-      _: undefined,
-      {
-        first,
-        last,
-        before,
-        after,
-        orderBy = 'id',
-        orderDirection = 'asc',
-      }: Pagination,
-      { db }: Context,
-    ): Promise<object[]> => {
-      if (first < 0) {
-        throw new UserInputError('First must be positive number');
-      }
-
-      if (last < 0) {
-        throw new UserInputError('Last must be positive number');
-      }
-
-      const baseQuery = db<Menu>('menus');
-
-      const data = await paginate(baseQuery, {
-        first: !first && !last ? 10 : first,
-        last,
-        before,
-        after,
-        orderBy,
-        orderDirection,
-      });
-
-      if (data) {
-        return normalizeDatabaseData(data);
-      }
-
-      return [];
-    },
-  },
-  Menu: {
-    restaurant: async (
-      { restaurantId }: { restaurantId: number },
-      _: undefined,
-      { db }: Context,
-    ): Promise<object | undefined> => {
-      const data = await db<Restaurant>('restaurants').where(
-        'id',
-        restaurantId,
-      );
-
-      if (data) {
-        return normalizeDatabaseData(data[0]);
-      }
-    },
-    menuItems: async (
-      { id }: { id: number },
-      _: undefined,
-      { db }: Context,
-    ): Promise<object | undefined> => {
-      const data = await db<MenuItem>('menu_items')
-        .where('menu_id', id)
-        .orderBy('weight');
-
-      if (data) {
-        return normalizeDatabaseData(data);
-      }
-    },
-  },
-};
