+++ src/menuItemComponent/resolvers.ts
@@ -1,0 +1,6 @@
+import { MenuItemComponentType } from './interfaces';
+import { generateEnumResolver } from '../utils/graphql';
+
+export default {
+  MenuItemComponentType: generateEnumResolver(MenuItemComponentType),
+};
